+++ packages/dates/src/utilities/formatDate.ts
@@ -1,19 +1,45 @@
-import {memoize} from '@shopify/function-enhancers';
+const intl = new Map();
+const memoizedGetDateTimeFormat = function(locale, options) {
+  const key = dateTimeFormatCacheKey(locale, options);
+  if (intl.has(key)) {
+    return intl.get(key);
+  }
+  const i = new Intl.DateTimeFormat(locale, options);
+  intl.set(key, i);
+  return i;
+};
 
-const memoizedGetDateTimeFormat = memoize(
-  getDateTimeFormat,
-  dateTimeFormatCacheKey,
-);
+const browserFeatureDetectionDate = Intl.DateTimeFormat('en', {
+  hour: 'numeric',
+});
+
+interface FormatDateOptions extends Intl.DateTimeFormatOptions {
+  hourCycle?: string;
+}
+
+const resolvedOptions: FormatDateOptions | undefined =
+  typeof browserFeatureDetectionDate.resolvedOptions === 'undefined'
+    ? undefined
+    : browserFeatureDetectionDate.resolvedOptions();
 
 export function formatDate(
   date: Date,
   locales: string | string[],
-  options: Intl.DateTimeFormatOptions = {},
+  options: FormatDateOptions = {},
 ) {
+  const hourCycleRequired =
+    resolvedOptions != null &&
+    options.hour12 === false &&
+    resolvedOptions.hourCycle != null;
+
+  if (hourCycleRequired) {
+    options.hour12 = undefined;
+    options.hourCycle = 'h23';
+  }
+
   // Etc/GMT+12 is not supported in most browsers and there is no equivalent fallback
   if (options.timeZone != null && options.timeZone === 'Etc/GMT+12') {
     const adjustedDate = new Date(date.valueOf() - 12 * 60 * 60 * 1000);
-
     return memoizedGetDateTimeFormat(locales, {
       ...options,
       timeZone: 'UTC',
@@ -21,13 +47,6 @@ export function formatDate(
   }
 
   return memoizedGetDateTimeFormat(locales, options).format(date);
-}
-
-function getDateTimeFormat(
-  locales?: string | string[],
-  options?: Intl.DateTimeFormatOptions,
-) {
-  return Intl.DateTimeFormat(locales, options);
 }
 
 function dateTimeFormatCacheKey(