+++ packages/react-shopify-app-route-propagator/src/index.ts
@@ -1,2 +1,0 @@
-export {default as RoutePropagator} from './route-propagator';
-export {default as useRoutePropagation} from './hook';